refactor(oui-radio): update project structure
