+++ packages/oui-angular/src/index.spec.js
@@ -25,8 +25,6 @@ loadTests(require.context("../../oui-pagination/src/", true, /.*((\.spec)|(index
 loadTests(require.context("../../oui-popover/src/", true, /.*((\.spec)|(index))$/));
 loadTests(require.context("../../oui-progress/src/", true, /.*((\.spec)|(index))$/));
 loadTests(require.context("../../oui-radio/src/", true, /.*((\.spec)|(index))$/));
-loadTests(require.context("../../oui-radio-group/src/", true, /.*((\.spec)|(index))$/));
-loadTests(require.context("../../oui-radio-toggle-group/src/", true, /.*((\.spec)|(index))$/));
 loadTests(require.context("../../oui-search/src/", true, /.*((\.spec)|(index))$/));
 loadTests(require.context("../../oui-select/src/", true, /.*((\.spec)|(index))$/));
 loadTests(require.context("../../oui-select-picker/src/", true, /.*((\.spec)|(index))$/));
