--- packages/oui-radio-group/src/index.js
@@ -1,9 +1,0 @@
-import "@oui-angular/oui-radio/src";
-import RadioGroup from "./radio-group.component.js";
-
-export default angular
-    .module("oui.radio-group", [
-        "oui.radio"
-    ])
-    .component("ouiRadioGroup", RadioGroup)
-    .name;
