+++ packages/oui-radio/src/index.js
@@ -1,6 +1,10 @@
-import Radio from "./radio.component.js";
+import Radio from "./radio.component";
+import RadioGroup from "./group/radio-group.component";
+import RadioToggleGroup from "./toggle-group/radio-toggle-group.component";
 
 export default angular
     .module("oui.radio", [])
     .component("ouiRadio", Radio)
+    .component("ouiRadioGroup", RadioGroup)
+    .component("ouiRadioToggleGroup", RadioToggleGroup)
     .name;
