+++ packages/oui-radio/src/toggle-group/radio-toggle-group.component.js
@@ -1,4 +1,4 @@
-import controller from "@oui-angular/oui-radio-group/src/radio-group.controller";
+import controller from "../group/radio-group.controller";
 
 export default {
     template: "<div ng-transclude class='oui-radio-toggle-group'></div>",
